+++ src/main/java/mods/belgabor/amtweaker/mods/ie/IE.java
@@ -1,0 +1,8 @@
+package mods.belgabor.amtweaker.mods.ie;
+
+/**
+ * Created by Belgabor on 19.06.2016.
+ */
+public class IE {
+    public static final String MODID = "ImmersiveEngineering"; 
+}