+++ src/main/java/mods/belgabor/amtweaker/mods/ss2/SS2.java
@@ -1,10 +1,17 @@
 package mods.belgabor.amtweaker.mods.ss2;
 
+import blusunrize.immersiveengineering.api.tool.ExternalHeaterHandler;
+import cpw.mods.fml.common.Loader;
 import minetweaker.MineTweakerAPI;
 import mods.belgabor.amtweaker.AMTweaker;
+import mods.belgabor.amtweaker.mods.ie.IE;
 import mods.belgabor.amtweaker.mods.ss2.handlers.*;
 import mods.belgabor.amtweaker.mods.ss2.util.SS2AccessHelper;
+import mods.belgabor.amtweaker.mods.ss2.xmod.IEHeaterAdapters;
 import org.apache.logging.log4j.Level;
+import shift.sextiarysector.tileentity.TileEntityFluidFurnace;
+import shift.sextiarysector.tileentity.TileEntityFoodSmokers;
+import shift.sextiarysector.tileentity.TileEntityLargeFurnace;
 
 /**
  * Created by Belgabor on 04.06.2016.
@@ -12,6 +19,9 @@ import org.apache.logging.log4j.Level;
 public class SS2 {
     public static final String MODID = "SextiarySector";
     public static boolean available = false;
+    
+    public static boolean configIEHeatFurnaces = true;
+    public static boolean configIEHeatSmokers = false;
     
     public SS2() {
         MineTweakerAPI.registerClass(Extractor.class);
@@ -35,6 +45,20 @@ public class SS2 {
             AMTweaker.log(Level.INFO, "Successfully gained access to SS2 classes.");
         } else {
             AMTweaker.log(Level.WARN, "Could not get access to SS2 classes. Support incomplete.");
+        }
+        
+        if (Loader.isModLoaded(IE.MODID) && (configIEHeatFurnaces || configIEHeatSmokers)) {
+            initIEIntegration();
+        }
+    }
+
+    private void initIEIntegration() {
+        if (configIEHeatFurnaces) {
+            ExternalHeaterHandler.registerHeatableAdapter(TileEntityFluidFurnace.class, new IEHeaterAdapters.LiquidFurnaceAdapter());
+            ExternalHeaterHandler.registerHeatableAdapter(TileEntityLargeFurnace.class, new IEHeaterAdapters.LargeFurnaceAdapter());
+        }
+        if (configIEHeatSmokers) {
+            ExternalHeaterHandler.registerHeatableAdapter(TileEntityFoodSmokers.class, new IEHeaterAdapters.LiquidFurnaceAdapter());
         }
     }
 }